# final-project/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple service for media social",
        "title": "MyGram API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/media-social": {
            "get": {
                "description": "Retrieve all media social entries",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media-social"
                ],
                "summary": "Get all media social entries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MediaSocialResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.TErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new media social entry with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media-social"
                ],
                "summary": "Create a new media social entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Media Social data",
                        "name": "mediaSocialRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MediaSocialRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.MediaSocialResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.TErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.TErrorResponse"
                        }
                    }
                }
            }
        },
        "/media-social/{medsos_id}": {
            "get": {
                "description": "Retrieve media social entry details by media social ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media-social"
                ],
                "summary": "Get media social entry by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Media Social ID",
                        "name": "medsos_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MediaSocialResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.TErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.TErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update media social entry details by media social ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media-social"
                ],
                "summary": "Update media social entry by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Media Social ID",
                        "name": "medsos_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Media Social details",
                        "name": "mediaSocialRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MediaSocialRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MediaSocialResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.TErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.TErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete media social entry by media social ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media-social"
                ],
                "summary": "Delete media social entry by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Media Social ID",
                        "name": "medsos_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.TSuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.TErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "request.MediaSocialRequest": {
            "type": "object",
            "properties": {
                "media_social_url": {
                    "type": "string",
                    "example": "https://www.instagram.com/fildzaanf"
                },
                "name": {
                    "type": "string",
                    "example": "instagram"
                }
            }
        },
...
        "response.MediaSocialResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "media_social_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "responses.TErrorResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "$ref": "#/definitions/responses.TResponseMeta"
                }
            }
        },
        "responses.TResponseMeta": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "responses.TSuccessResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "$ref": "#/definitions/responses.TResponseMeta"
                },
                "results": {}
            }
        }
    }
}